Modified: src

+++ src/Main.java
@@ -2,7 +2,7 @@ public class Main {
     public static void main (String[] args) {
         int ticket = 4023;
         int part_ticket = 20;
-        int total_miles = (int) ticket / (int)part_ticket;
+        int total_miles = ticket / part_ticket;
         System.out.println(total_miles);
     }
 }
